feat: Refactor Commenter into action-utils
