+++ package.json
@@ -15,7 +15,6 @@
   },
   "devDependencies": {
     "@appland/appmap-agent-js": "^13.7.0",
-    "@octokit/rest": "^20.0.1",
     "@types/argparse": "^2.0.10",
     "@types/glob": "^8.1.0",
     "@types/jest": "^29.5.0",
@@ -33,11 +32,10 @@
   },
   "dependencies": {
     "@actions/artifact": "^1.1.1",
-    "@actions/core": "^1.10.0",
-    "@actions/github": "^5.1.1",
-    "@appland/action-utils": "https://github.com/getappmap/action-utils",
+    "@appland/action-utils": "https://github.com/getappmap/action-utils#19e6956b6b078a95154d0ba3fea2c762060b0b45",
     "argparse": "^2.0.1",
-    "handlebars": "^4.7.7"
+    "handlebars": "^4.7.7",
+    "openapi-types": "^12.1.3"
   },
   "engines": {
     "node": ">=18.0.0"
