--- test/util.ts
@@ -1,4 +1,0 @@
-import path from "path";
-
-export const dataDir = path.join(__dirname, 'data');
-export const reportPath = path.join(dataDir, 'report.md');
